Rename demos/default/client/fake_fetch.ts to http/common/fake_fetcher.ts
