+++ http/common/fake_fetcher.ts
@@ -1,5 +1,5 @@
-import type { DefaultData } from "../../../http/common/default_data.ts";
-import type { Fetcher } from "../../../http/common/fetcher.ts";
+import type { DefaultData } from "./default_data.ts";
+import type { Fetcher } from "./fetcher.ts";
 
 interface FakeData extends DefaultData {
   $key: "id";
@@ -8,7 +8,7 @@ interface FakeData extends DefaultData {
 /**
  * FakeFetch is a fake fetch function that simulates server responses on the client.
  */
-export class FakeFetch implements Fetcher {
+export class FakeFetcher implements Fetcher {
   constructor(
     public data: Record<string, FakeData> = {},
   ) {}
